+++ ChangeLog.rst
@@ -1,3 +1,8 @@
+HEAD
+====
+ - A new class :class:`modelcif.ReferenceDatabase` allows describing
+   collections of sequences that were used as part of the modeling protocol.
+
 0.4 - 2022-04-14
 ================
  - Sequence information for templates is now only written to template-specific